# [1/2] Add concat_mla_absorb_k for flashinfer mla to speedup FlashInferMLA