@yuan-luo 2025-09-28
> Please add gsm8k acc


gsm8k result, no acc drop.
```
➜  python git:(main) ✗ python3 -m sglang.launch_server --model-path deepseek-ai/DeepSeek-R1 --host 0.0.0.0 --port 30000 --trust-remote-code --enable-cache-report --quantization fp8 --log-level info --max-running-requests 16 --mem-fraction-static 0.92 --chunked-prefill-size 16384 --context-length 65535 --tp-size 8 --enable-metrics --cuda-graph-max-bs 16 --disable-radix-cache --page-size 64
```

```
This PR:
➜  sglang_dev git:(refactor_concat_mla_absorb) ✗ python3 benchmark/gsm8k/bench_sglang.py --num-questions 200 --parallel 128 --num-shots 8 --port 30000
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 200/200 [00:53<00:00,  3.76it/s]
Accuracy: 0.960
Invalid: 0.000
Latency: 53.431 s
Output throughput: 380.659 token/s


Main:
➜  sglang_dev2 git:(main) ✗ python3 benchmark/gsm8k/bench_sglang.py --num-questions 200 --parallel 128 --num-shots 8 --port 30000
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 200/200 [00:54<00:00,  3.68it/s]
Accuracy: 0.960
Invalid: 0.000
Latency: 54.601 s
Output throughput: 367.944 token/s

```